+++ client/src/app/app-shell/header/header.component.scss
