+++ client/src/app/app-shell/mode-toggle/mode-toggle.component.scss
