+++ client/src/app/dashboard/active-list/task/task.component.scss
@@ -59,8 +59,8 @@
       &::after {
         content: '';
         position: absolute;
-        background: center / contain url('../../assets/images/icon-check.svg')
-          no-repeat;
+        background: center / contain
+          url('../../../../assets/images/icon-check.svg') no-repeat;
         inset: 0;
         height: 0.625rem;
         width: 0.625rem;
