+++ client/src/app/dashboard/dashboard.routing.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { DashboardComponent } from './dashboard/dashboard.component';
+import { DashboardComponent } from './dashboard.component';
 
 const DASHBOARD_ROUTES: Routes = [
   {
