+++ client/src/app/landing/auth.guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from './auth.service';
 
 export const authGuard: CanActivateFn = async () => {
   const authService = inject(AuthService);
