+++ client/src/app/landing/landing.component.ts
@@ -6,7 +6,7 @@ import {
   signal,
 } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from './auth.service';
 import { NgOptimizedImage } from '@angular/common';
 import { ClipboardModule } from '@angular/cdk/clipboard';
