+++ client/src/app/shared/new-task/new-task.component.scss
